agent/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
google-adk
google-genai
ag-ui-adk
composio>=0.5.0
httpx
